[custom-exception] Setup Custom Exceptions

- Well Done!!!

+++ src/Core/HR.LeaveManager.Application/Feature/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommaindHandler.cs
@@ -1,4 +1,6 @@
-﻿using HR.LeaveManager.Application.Contracts.Persistence;
+﻿using HR.LeaveManagement.Domain;
+using HR.LeaveManager.Application.Contracts.Exceptions;
+using HR.LeaveManager.Application.Contracts.Persistence;
 using MediatR;
 
 namespace HR.LeaveManager.Application.Feature.LeaveType.Commands.DeleteLeaveType;
@@ -15,6 +17,9 @@ public class DeleteLeaveTypeCommaindHandler : IRequestHandler<DeleteLeaveTypeCom
 	public async Task<Unit> Handle(DeleteLeaveTypeCommaind request, CancellationToken cancellationToken)
 	{
 		var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+		if (leaveTypeToDelete is null)
+			throw new NotFoundException(nameof(LeaveType), request.Id);
 
 		await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
